Resolved issues with package installation
CLOPLO_Mapping pip installs were in separate cells and was updated to be in a single cell using %%bash. Also included pip install pandas since this was not included. Classification_Bloom_SOLO_Mapping included pip install nltk as this was not included. Also included code to run setup.py and install dependencies for docx2csv.